Set DisplayArea in DesktopModeLaunchParamsModifier
When desktop windowing is enabled, DesktopModeLaunchParamsModifier is applied to any new launching activity, and for any non-null task it superceeds TaskLaunchParamsModifier. Usually TaskLaunchParamsModifier sets the preferredDisplayArea, but since it doesn't, the launchParams for tasks have null display areas. This was leading to an issue where taskbar tests were failing because they were launching an app in split screen unexpectedly. Usually, if there is display area in the launch params, the task is reparented to the display area and inherits the windowing mode of the display area.

Flag: None
Bug: 335856622
Bug: 333446588
Test: atest NexusLauncherTests:com.android.quickstep
Change-Id: Ifb347a6c93b94f7c197fbe722db06200884442df