Use KeyProperties.UID_SELF instead of KeyStore.UID_SELF
Since the class 'android.security.KeyStore' is going away, the UID_SELF
constant is being moved to a more appropriate location in KeyProperties.
Update code that references this constant to use the new location.  No
change in behavior; the value of the constant remains the same.

Bug: 326508120
Test: build
Flag: NONE mechanical refactoring with no behavior change
Change-Id: I86aa1add526a9eb45bb89a73cc18bb54ba3496c3