Fix comparison of route id with original route id
This change is a partial revert of ag/25865380. ROUTE_ID_DEFAULT is an
original route id, as opposed to what MediaRoute2Info#getId returns,
which is a unique id (original id plus a disambiguating prefix). So we
are replacing the use of ROUTE_ID_DEFAULT with defaultRoute#getId() to
compare among unique route ids, instead of comparing a unique id with a
original id.

Bug: b/319645714
Test: atest CtsMediaHostTestCases CtsMediaBetterTogetherTestCases
Flag: EXEMPT bugfix
Change-Id: Ide0e0e3c46849ecc4e634af038279067d5d3c5d8