Simplify IMMS.MyPackageMonitor
With my previous changes [1][2], long and/or frequent operations in

 MyPackageMonitor#onFinishPackageChangesInternal()

are expected to be no longer contributing to UI janks. Now it is time
to focus more on the correctness rather than trying to optimize things
in a complicated way. Complicated optimizations make it difficult for
us to diagnose issues like Bug 335281466.

This CL basically replaces my past optimizations [3][4] in a more
strict and straightforward way. We first create a new InputMethodMap
based on the latest results returned from the PackageManagerService,
then see if there is any actual data change from the current
InputMethodMap. If those two InputMethodMap instances are exactly the
same, it means that no further update is necessary.

There should be no semantic observable change.

 [1]: I263cd49dd4d64b64136acc3dad469f83a862ce97
      51ef90c063d51e3febe78fb89987cf9f93bbdc90
 [2]: Icc84fe9d0d3b5842b4d5ee51b12b0d3cd10a1d24
      cc57a48664d9857aa93455fdf4f6d41fe69f0a73
 [3]: I7b69c349318ce06a48d03a4468cf2c45bfb73dc2
      d2ddf2ed84abb4b3f6511029c852c8a16eec6a41
 [4]: I063688297156188f68fe0b55a46d72f2e811dc88
      1091b849b2359d7686e69822ae6184ec0ee5a32c

Bug: 329703038
Test: presubmit
Change-Id: I6e9de6ff5c6dd55e3c40f23586110cb713c62851