Glitchy notifs on clock size change
Notifications weren't animating Y position changes as expected. In order
to synchronize the animation, we tether the intrablueprint transition
state into KeyguardRootViewBinder which listens for placeholder position
changes and forwards them to the notification code. This function now
skips certain layouts during transitions to prevent later layout changes
from conflicting with the running animation. These layouts are erronous
to the transition and suppressed by the transition directly for other
relevant lockscreen elements.

To facilitate two binders relying on the same transition state, I've
moved the transition management logic from the binder to the model, but
otherwise not modified it much. This makes the state available to more
components, allows the binder to be stateless, and lets us convert it
to being a top-level kotlin object like other binders.

I've also corrected an issue where the notif shelf would animate in when
the clock size changed. AodNotificationIconsSection.applyConstraints was
not setting the visibility, so it'd default to VISIBLE for a moment when
it was applied. The transition would pick up this incorrect transient
visibility value and animate the element in.

Bug: 341932557
Test: Manual & Presubmits
Flag: com.android.systemui.migrate_clocks_to_blueprint
Change-Id: Ib447caeb142f55a04f77f34cf8bd7fe6e7e83896