Don't call setLockScreenShown until we know if it's showing.
Early calls to setLockscreenShown with aodVisible=true/false were causing us to apply the default lockscreenShowing=true value, which causes problems if we boot with the lockscreen disabled.

The new KeyguardTransitionBootInteractor guarantees that we'll be provided a lockscreenShowing=true/false value soon after boot, so we can just wait for that.

Test: atest WindowManagerLockscreenVisibilityManagerTest
Flag: EXEMPT bugfix
Bug: 278086361
Change-Id: I50b3bd44a10efb1a7433af5a2468e83eba8f975a