Fix concurrentModificationException of DevicePostureController listeners
Some listener ended up registering or unregistering other listeners while posture updates were sent. I coulnd't repro it, but we had a few reports from the field. I suspect there is a chain of calls that leads to the destruction of some objects (onDestroy) and triggers the postureController.removeListener, but only rarely.

Fix: 345390663
Test: DevicePostureControllerImplTest
Flag: NONE safe fix for rare and not reproducible exception
Change-Id: I00b5813267e917c65ca269324a9f17783069fd82