Remove getCurrentInputMethodSubtypeForNonCurrentUsers()
This is a follow up CL to my previous CL [1], which enabled background
users to call

  InputMethodManager#getCurrentInputMethodSubtype()

by introducing

  #getCurrentInputMethodSubtypeForNonCurrentUsers().

The reason why we could not simply reuse the same implementation for

  InputMethodManagerService#getCurrentInputMethodSubtypeLocked()

was due to the fact that IMMS#mCurrentSubtype was not per-user.

Now that #mCurrentSubtype is fully per-user [2], we should be able to
unify the code.

 [1]: Icb09f9cb1a4147884faa9952b2e03ec4afa9f0b1
      01437218915789432c5ca1f295b55ca45c469d37
 [2]: Ibcc39212ead12c2897f8f52cb9f15b1d39959f71
      5c2c33fc82bc0ac0cb158b910628b450318ccc47

Bug: 325515685
Bug: 350386877
Test: presubmit
Flag: EXEMPT refactor
Change-Id: Ia52b556ff3df3d2b50b9d62198455e5ce6ebe659