Fix black veil when splitting from side in some conditions
This CL explicitly sets alpha to 0 or 1 when SplitDecorManager does an immediate visibility change.

Previously, the layer visibility was being set, but not the alpha, so in some situations the alpha was not updated properly and the veil was not visible during the transition. Setting the alpha explicitly fixes the problem.

Fixes: 343859372
Test: Manually verified that the veil is visible in the buggy repro condition.
Flag: EXEMPT bugfix
Change-Id: I439d478ad390da301b7c8faca97b272b7e727a24
(cherry picked from commit a7f3818d7a14e75d6d0ea2f7b4dba5fbc108c2b0)