Remove single log extraction from  log_list.json
Instead of writing a file for each log in the log
list, going forward we want to keep the whole json
file as-is, and let Conscrypt handle the process
of parsing the json data.

I did keep the step where the log list is read
into memory and passed into a JSONObject, to give
Conscrypt a guarantee that we are at least giving
them something that looks like a json file.

Flag: com.android.server.updates.certificate_transparency_installer
Test: am broadcast -a com.google.android.configupdater.CTLogs.UPDATE_CT_LOGS -n com.google.android.configupdater/.CTLogs.CTLogsUpdateRequestReceiver and ls /data/misc/keychain/ct/
Bug: 319829948
Change-Id: If90ca577063a4bd2216a470e5b14d65a2dffc5ab