Apply font weight adjustment to cached spannables if needed
Problem: When creating a spannable in StringBlock from html,
fontWeightAdjustment is applied for the <b> tag.
This instance is then cached. On returning the cached instance
when the page is opened again, the text is not updated with
the new fontWeightAdjustment.

Solution: We cache the weight adjustment, check for bold
spans in the cached instance, and apply a new StyleSpan
with the font weight adjustment if the adjustments are different.

Test: manual
Bug: 379962518
Flag: android.content.res.apply_font_weight_adjustment_on_cached_strings

Change-Id: I732147918ba36913ff89d90676e71cf8ea78d579