Create the EventReceiver with the right display Id
The current context will almost always have the default display, we need
to create the EventReceiver with the displayId of the display it is
handling.

For testing, open a freeform window on a secondary display with the dev
option `Force Desktop Mode` (this registers an EventReceiver) and try
drag the app handle.

Bug: 352316177
Flag: EXEMPT bug fix
Test: Manual (see description)
Change-Id: I0f2ac32996bf356c8840d4974b98d70967aa00fb