Run IMMS#onUnlockUser() also for background users
As part of our on-going effort to address the remaining dependency on

  InputMethodManagerService#mCurrentUserId,

this CL updates the behavior of

  InputMethodManagerService#onUnlockUser()

so that it calls

  * IMMS#postInputMethodSettingUpdatedLocked()
  * IMMS#updateInputMethodsFromSettingsLocked()

not only for the current user but also for background users, even for
phones/tablets/TVs/wearables. This was not possible until when we
update the above two methods to take userId [1].

With this CL, we would start initializing IME-related settings
(e.g. secure settings) more eagerly for background users. But other
than that, user observable behavior should remain unchanged.

 [1]: I9d4615e6da979204c21555e040a687163c1460ab
      084edd67dfada6b0d58850327665869a0249910c

Bug: 350386877
Test: presubmit
Flag: android.view.inputmethod.concurrent_input_methods
Change-Id: Iecbca2696d4e2fd56a434f2e3a95d14976978e06