Support force-stop IME package for background users
This is a follow CL to my previous CL [1], which restored a certain
de-facto behavior when triggering force-stop against a package that
hosts IMEs.

As explained in the commit message [1], triggering force-stop against
a background user's IME package has been no-op since the initial
multi-user support was introduced for IMMS [2].

With my another CL [3], however, we are finally ready to propagate
force-stop signal for a background user to relevant downstream code.

Let's remove this inconsistency as it is really not intuitive.

There is also a corresponding CTS change [4].

 [1]: I261e1c60050ee20f7673425f75588269f865ba50
      a0f9bd5332a3143ac63ced5b33841430b857d9d5
 [2]: Ib23849d352db33f0747aa9d5a178f00ac726c13b
      b67143b01df1c4573ff5145cf4eec04c5098e85d
 [3]: I0010ab8c930c5d697227ef107264dadaea398881
      889b0c105819984dfc2e73c841663452bf1c3a52
 [4]: Ia19d5a578297eaf9dcd344822352feda71781bb8

Bug: 350386877
Fix: 352758479
Test: atest CtsInputMethodInstallTestCases:ForceStopTest
Flag: android.view.inputmethod.concurrent_input_methods
Change-Id: Ib953d7968ea3170865f82508d616ffc4e8e91d7c